requirements.txt
# for additional pywaggle install options, see: https://github.com/waggle-sensor/pywaggle#installation-guides
pywaggle==0.56.0
pyserial==3.5

